GetNewsAPI/requirements.txt
flask
mysql-connector-python
apscheduler
requests